#1872: fix header generation; move documentView files to folder
